Split up run/help commands into separate packages
