+++ console-runner/src/main/java/info/javaspec/console/Command.java
@@ -1,6 +1,6 @@
 package info.javaspec.console;
 
 @FunctionalInterface
-interface Command {
+public interface Command {
   int run();
 }
