+++ console-runner/src/main/java/info/javaspec/console/Reporter.java
@@ -1,6 +1,7 @@
 package info.javaspec.console;
 
 import info.javaspec.RunObserver;
+import info.javaspec.console.help.HelpObserver;
 
 /** An observer of all things, that handles reporting when running commands. */
 public interface Reporter extends HelpObserver, RunObserver { }
