+++ console-runner/src/main/java/info/javaspec/console/help/HelpCommand.java
@@ -1,8 +1,10 @@
-package info.javaspec.console;
+package info.javaspec.console.help;
+
+import info.javaspec.console.Command;
 
 import java.util.Arrays;
 
-final class HelpCommand implements Command {
+public final class HelpCommand implements Command {
   private final HelpObserver observer;
 
   public HelpCommand(HelpObserver observer) {
