+++ console-runner/src/test/java/info/javaspec/console/StaticCommandFactoryTest.java
@@ -3,6 +3,10 @@ package info.javaspec.console;
 import de.bechte.junit.runners.context.HierarchicalContextRunner;
 import info.javaspec.RunObserver;
 import info.javaspec.console.ArgumentParser.CommandFactory;
+import info.javaspec.console.help.DetailedHelpCommand;
+import info.javaspec.console.help.HelpCommand;
+import info.javaspec.console.help.HelpObserver;
+import info.javaspec.lang.lambda.RunSpecsCommand;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
