+++ console-runner/src/test/java/info/javaspec/console/help/HelpCommandTest.java
@@ -1,6 +1,9 @@
-package info.javaspec.console;
+package info.javaspec.console.help;
 
 import de.bechte.junit.runners.context.HierarchicalContextRunner;
+import info.javaspec.console.Command;
+import info.javaspec.console.help.HelpCommand;
+import info.javaspec.console.help.HelpObserver;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
